fully working (hopefully) side config with GUI

+++ src/main/java/io/enderdev/linkedtanks/tiles/TileLinkedTank.kt
@@ -7,26 +7,30 @@ import io.enderdev.linkedtanks.data.LTPersistentData
 import io.enderdev.linkedtanks.tiles.buttons.DeleteButtonWrapper
 import io.enderdev.linkedtanks.tiles.buttons.LinkButtonWrapper
 import io.enderdev.linkedtanks.tiles.buttons.RenameButtonWrapper
+import io.enderdev.linkedtanks.tiles.buttons.SideConfigurationButtonWrapper
 import io.enderdev.linkedtanks.tiles.util.FluidSideConfiguration
 import io.enderdev.linkedtanks.util.LinkedFluidHandler
 import io.enderdev.linkedtanks.util.extensions.dim
 import io.enderdev.linkedtanks.util.extensions.dimId
+import net.minecraft.entity.player.EntityPlayer
 import net.minecraft.nbt.NBTTagCompound
+import net.minecraft.network.NetworkManager
+import net.minecraft.network.play.server.SPacketUpdateTileEntity
 import net.minecraft.util.EnumFacing
 import net.minecraft.util.ITickable
 import net.minecraftforge.common.capabilities.Capability
 import net.minecraftforge.fluids.FluidRegistry
-import net.minecraftforge.fml.common.network.simpleimpl.MessageContext
 import org.ender_development.catalyx.client.button.AbstractButtonWrapper
 import org.ender_development.catalyx.client.button.PauseButtonWrapper
 import org.ender_development.catalyx.client.button.RedstoneButtonWrapper
 import org.ender_development.catalyx.client.gui.BaseGuiTyped
 import org.ender_development.catalyx.tiles.BaseTile
 import org.ender_development.catalyx.tiles.helper.IButtonTile
+import org.ender_development.catalyx.tiles.helper.ICopyPasteExtraTile
 import org.ender_development.catalyx.tiles.helper.IFluidTile
 import org.ender_development.catalyx.tiles.helper.IGuiTile
 
-class TileLinkedTank : BaseTile(LinkedTanks.modSettings), IFluidTile, ITickable, IGuiTile, IButtonTile, BaseGuiTyped.IDefaultButtonVariables {
+class TileLinkedTank : BaseTile(LinkedTanks.modSettings), IFluidTile, ITickable, IGuiTile, IButtonTile, BaseGuiTyped.IDefaultButtonVariables, ICopyPasteExtraTile {
 	override var isPaused = false
 	override var needsRedstonePower = false
 
@@ -41,12 +45,27 @@ class TileLinkedTank : BaseTile(LinkedTanks.modSettings), IFluidTile, ITickable,
 	override val fluidHandler = LinkedFluidHandler(null)
 
 	var channelUpdateTicks = 2
+	var channelUnfuckTicks = 10
 	override fun update() {
-		markDirtyGUIEvery(7)
+		// don't tick on client-side
+		if(world.isRemote)
+			return
+
+		markDirtyGUIEvery(5)
 		if(++channelUpdateTicks == 3) {
 			channelUpdateTicks = 0
 			if((channelId != Constants.NO_CHANNEL && channelData == null) || channelData?.deleted == true)
 				unlink()
+
+			// this shouldn't happen but might as well
+			if(++channelUnfuckTicks == 20) {
+				channelUnfuckTicks = 0
+				channelData?.let {
+					val pos = pos dim world.dimId
+					if(it.linkedPositions.add(pos))
+						LinkedTanks.logger.debug("Channel id {} didn't have us ({}) added to linked positions! This shouldn't happen!", channelId, pos)
+				}
+			}
 		}
 		fluidSideConfiguration.tick()
 	}
@@ -64,11 +83,9 @@ class TileLinkedTank : BaseTile(LinkedTanks.modSettings), IFluidTile, ITickable,
 		channelData = null
 	}
 
-	fun link(newChannelId: Int, ctx: MessageContext) {
+	fun link(newChannelId: Int, player: EntityPlayer) {
 		if(channelId == newChannelId)
 			return
-
-		val player = ctx.serverHandler.player
 
 		// unlink
 		if(newChannelId == Constants.NO_CHANNEL) {
@@ -110,7 +127,8 @@ class TileLinkedTank : BaseTile(LinkedTanks.modSettings), IFluidTile, ITickable,
 		if(newChannelId == Constants.NO_CHANNEL)
 			return
 
-		newChannelData.linkedPositions.add(pos dim world.dimId)
+		if(world != null) // in early readFromNBT, world is still null
+			newChannelData.linkedPositions.add(pos dim world.dimId)
 		channelId = newChannelId
 		channelData = newChannelData
 
@@ -122,16 +140,17 @@ class TileLinkedTank : BaseTile(LinkedTanks.modSettings), IFluidTile, ITickable,
 			is PauseButtonWrapper -> isPaused = !isPaused
 			is RedstoneButtonWrapper -> needsRedstonePower = !needsRedstonePower
 			is LinkButtonWrapper -> {
+				val player = button.ctx.serverHandler.player
 				val newChannelId = if(button.channelId == Constants.CREATE_NEW_CHANNEL)
-					LTPersistentData.createNewChannel(button.ctx.serverHandler.player, this)
+					LTPersistentData.createNewChannel(player, this)
 				else
 					button.channelId
 
-				link(newChannelId, button.ctx)
+				link(newChannelId, player)
 			}
 			is RenameButtonWrapper -> {
 				channelData?.let { channelData ->
-					if(button.ctx.serverHandler.player.uniqueID != channelData.ownerUUID)
+					if(!channelData.canBeEditedBy(button.ctx.serverHandler.player.uniqueID))
 						return
 
 					var newName = button.newName.trim()
@@ -147,13 +166,24 @@ class TileLinkedTank : BaseTile(LinkedTanks.modSettings), IFluidTile, ITickable,
 			}
 			is DeleteButtonWrapper -> {
 				channelData?.let { channelData ->
-					if(button.ctx.serverHandler.player.uniqueID != channelData.ownerUUID)
+					if(!channelData.canBeEditedBy(button.ctx.serverHandler.player.uniqueID))
 						return
 
 					channelData.deleted = true
 					unlink()
 					markDirtyGUI()
 				}
+			}
+			is SideConfigurationButtonWrapper -> {
+				if(channelData?.canBeEditedBy(button.ctx.serverHandler.player.uniqueID) == false)
+					return
+
+				if(button.affectsAll)
+					fluidSideConfiguration.sides.replaceAll { _, _ -> button.side }
+				else
+					fluidSideConfiguration.sides[button.facing] = button.side
+
+				markDirtyGUI()
 			}
 		}
 	}
@@ -162,6 +192,7 @@ class TileLinkedTank : BaseTile(LinkedTanks.modSettings), IFluidTile, ITickable,
 	override fun readFromNBT(compound: NBTTagCompound) {
 		if(compound.hasKey("ChannelId"))
 			link(compound.getInteger("ChannelId"))
+
 		fluidSideConfiguration.readFromNBT(compound.getCompoundTag("FluidSideConfiguration"))
 		super.readFromNBT(compound)
 	}
@@ -169,28 +200,41 @@ class TileLinkedTank : BaseTile(LinkedTanks.modSettings), IFluidTile, ITickable,
 	// server-side
 	override fun writeToNBT(compound: NBTTagCompound): NBTTagCompound {
 		super.writeToNBT(compound)
+
 		if(channelId != Constants.NO_CHANNEL)
 			compound.setInteger("ChannelId", channelId)
-		compound.setTag("FluidSideConfiguration", fluidSideConfiguration.writeToNBT())
+
+		compound.setTag("FluidSideConfiguration", fluidSideConfiguration.writeToNBT(false))
+
 		return compound
 	}
+
+	override fun getUpdatePacket() =
+		SPacketUpdateTileEntity(pos, 0, updateTag)
 
 	// server-side, sent to client-side [handleUpdateTag]
 	override fun getUpdateTag(): NBTTagCompound {
 		return NBTTagCompound().apply {
 			writeInternal(this) // write stuff like x,y,z + Forge stuff, since we don't call writeToNBT
-			setTag("FluidSideConfiguration", fluidSideConfiguration.writeToNBT())
+			setTag("FluidSideConfiguration", fluidSideConfiguration.writeToNBT(true))
 			setInteger("ChannelId", channelId)
+
 			channelData?.let { channelData ->
 				setString("Name", channelData.name)
 				setString("OwnerUsername", channelData.ownerUsername)
 				setUniqueId("OwnerUUID", channelData.ownerUUID)
+
 				if(channelData.fluid != null)
 					setString("FluidName", FluidRegistry.getFluidName(channelData.fluid))
+
 				setInteger("FluidAmount", channelData.fluidAmount)
 				setInteger("FluidCapacity", channelData.fluidCapacity)
 			}
 		}
+	}
+
+	override fun onDataPacket(net: NetworkManager, pkt: SPacketUpdateTileEntity) {
+		handleUpdateTag(pkt.nbtCompound)
 	}
 
 	// client-side, handle from server-side [getUpdateTag]
@@ -198,16 +242,18 @@ class TileLinkedTank : BaseTile(LinkedTanks.modSettings), IFluidTile, ITickable,
 		super.readFromNBT(tag)
 
 		fluidSideConfiguration.readFromNBT(tag.getCompoundTag("FluidSideConfiguration"))
+		channelId = tag.getInteger("ChannelId")
 
-		if(!tag.hasKey("OwnerUsername"))
+		if(!tag.hasKey("OwnerUsername")) {
+			channelData = null
 			return
+		}
 
 		val name = tag.getString("Name")
 		val ownerUsername = tag.getString("OwnerUsername")
 		val ownerUUID = tag.getUniqueId("OwnerUUID")!!
 		val fluid = FluidRegistry.getFluid(tag.getString("FluidName"))
 		val fluidAmount = tag.getInteger("FluidAmount")
-		channelId = tag.getInteger("ChannelId")
 		channelData = ChannelData(false, ownerUUID, ownerUsername, name, fluid, fluidAmount, Constants.NO_LINKED_POSITIONS).apply {
 			fluidCapacityOverride = tag.getInteger("FluidCapacity")
 		}
@@ -223,9 +269,30 @@ class TileLinkedTank : BaseTile(LinkedTanks.modSettings), IFluidTile, ITickable,
 			null
 	}
 
+	// ICopyPasteExtraTile
+	override fun copyData(tag: NBTTagCompound) {
+		tag.setInteger("ChannelId", channelId)
+		tag.setTag("FluidSideConfiguration", fluidSideConfiguration.writeToNBT(true))
+	}
+
+	override fun pasteData(tag: NBTTagCompound, player: EntityPlayer) {
+		if(channelData?.canBeEditedBy(player.uniqueID) == false)
+			return
+
+		if(tag.hasKey("ChannelId")) {
+			val channelId = tag.getInteger("ChannelId")
+			if(channelId != Constants.CREATE_NEW_CHANNEL) // sanity check
+				link(channelId, player)
+		}
+
+		if(tag.hasKey("FluidSideConfiguration"))
+			fluidSideConfiguration.readFromNBT(tag.getCompoundTag("FluidSideConfiguration"))
+	}
+
 	init {
 		AbstractButtonWrapper.registerWrapper(LinkButtonWrapper::class.java)
 		AbstractButtonWrapper.registerWrapper(RenameButtonWrapper::class.java)
 		AbstractButtonWrapper.registerWrapper(DeleteButtonWrapper::class.java)
+		AbstractButtonWrapper.registerWrapper(SideConfigurationButtonWrapper::class.java)
 	}
 }
